revamp muworker-socket
