+++ muworker-socket/package.json
@@ -2,7 +2,6 @@
   "name": "muworker-socket",
   "version": "0.0.0",
   "description": "Web Worker made available to mudb",
-  "main": "index.js",
   "scripts": {
     "lint": "tslint -c tslint.json 'src/**/*.ts'",
     "test": "browserify test/*.js | tape-run"
@@ -13,7 +12,9 @@
   },
   "keywords": [
     "mudb",
-    "network"
+    "network",
+    "multiplayer",
+    "game"
   ],
   "author": "He Diyi",
   "license": "MIT",
